Former-commit-id: 83c3b465de542a0c70379990c41495d59d930c21

+++ pqueens/schedulers/local_scheduler.py
@@ -1,4 +1,7 @@
+""" There should be a docstring """
+
 import os
+import sys
 from .scheduler import Scheduler
 
 
@@ -32,9 +35,8 @@ class LocalScheduler(Scheduler):
 
         return cls(base_settings, scheduler_name)
 
-
-# ----------------- ABSTRACT-METHODS THAT MUST BE IMPLEMENTED -----------------
-    def alive(self, process_id): # TODO: ok for now (gets called in resources)
+# ------------------- CHILD METHODS THAT MUST BE IMPLEMENTED ------------------
+    def alive(self, process_id):  # TODO: ok for now (gets called in resources)
         """ Check whether or not job is still running
 
         Args:
@@ -43,10 +45,30 @@ class LocalScheduler(Scheduler):
         Returns:
             bool: indicator if job is still alive
         """
-        try:
-            # Send an alive signal to proc
-            os.kill(process_id, 0)
-        except OSError:
+        alive = False
+        command_list = ['ps h -p', str(process_id)]
+        command_string = ' '.join(command_list)
+        stdout, _, p = super().run_subprocess(command_string)
+
+        if stdout:
+            sys.stderr.write("Job %d waiting in queue.\n" % (process_id))
+            alive = True
+        else:
+            sys.stderr.write("Job %d is held or suspended.\n" % (process_id))
+            alive = False
+
+        if not alive:
+            try:
+                # try to kill the job.
+                os.kill(process_id, 0)
+                sys.stderr.write("Killed job %d.\n" % (process_id))
+            except ValueError:
+                sys.stderr.write("Failed to kill job %d.\n" % (process_id))
+
             return False
         else:
             return True
+
+    def get_process_id_from_output(self):
+        """ docstring """
+        pass
